Rename onepasswordImporters to onepassword
Rename the folder
Fix all the imports
Remove onepasswordImporters from white-list
