+++ libs/common/src/services/import.service.ts
@@ -45,10 +45,10 @@ import { MeldiumCsvImporter } from "../importers/meldium-csv-importer";
 import { MSecureCsvImporter } from "../importers/msecure-csv-importer";
 import { MykiCsvImporter } from "../importers/myki-csv-importer";
 import { NordPassCsvImporter } from "../importers/nordpass-csv-importer";
-import { OnePassword1PifImporter } from "../importers/onepasswordImporters/onepassword-1pif-importer";
-import { OnePassword1PuxImporter } from "../importers/onepasswordImporters/onepassword-1pux-importer";
-import { OnePasswordMacCsvImporter } from "../importers/onepasswordImporters/onepassword-mac-csv-importer";
-import { OnePasswordWinCsvImporter } from "../importers/onepasswordImporters/onepassword-win-csv-importer";
+import { OnePassword1PifImporter } from "../importers/onepassword/onepassword-1pif-importer";
+import { OnePassword1PuxImporter } from "../importers/onepassword/onepassword-1pux-importer";
+import { OnePasswordMacCsvImporter } from "../importers/onepassword/onepassword-mac-csv-importer";
+import { OnePasswordWinCsvImporter } from "../importers/onepassword/onepassword-win-csv-importer";
 import { PadlockCsvImporter } from "../importers/padlock-csv-importer";
 import { PassKeepCsvImporter } from "../importers/passkeep-csv-importer";
 import { PassmanJsonImporter } from "../importers/passman-json-importer";
